Make mouse wheel work even with low sensitivity
A mouse wheel event has a float value to indicate the "speed". This
value is passed to Lua as an integer. In the case of very low mouse
wheel sensitivities, the speed value can be below 1, which got converted
to zero for the lua side. This effectively disabled the mouse wheel.

The problem does not exist with X11 because the mouse wheel speed is an
integer - at least as provided by SDL3.

The fix is to pass the float value to Lua instead.